smol changes

+++ src/enginelib/include/enginelib/search.h
@@ -1,8 +1,9 @@
 #ifndef _ENGINELIB_SEARCH_H_
 #define _ENGINELIB_SEARCH_H_
 
-#include "containerslib/heap.h"
+#include <stdio.h>
 
+#include "containerslib/heap.h"
 #include "page_indexerlib/index.h"
 
 typedef struct Search Search;
@@ -13,7 +14,7 @@ struct Search
     Heap *heap_pr_page; // Heap<int, string>
 };
 
-size_t enginelib_search(Index *index, /* TODO: PR data,*/ FILE *in, Search *out);
+long long int enginelib_search(Index *index, /* TODO: PR data,*/ FILE *in, Search *out);
 
 Heap *enginelib_search_get_ranking(Search *search_result);
